Specify spacing on periodic_tasks in manager.py
Some methods in compute/manager.py were using the @periodic_task
decorator to ensure they were called regularly, but were not passing
in the `spacing` kwarg.  In this case the method can only be called at
the default rate, calculated as min(60 seconds, <spacings from all
periodic_tasks>), regardless of the poll interval set in config.

Although there was throttling code which prevented the methods from
executing too often, the case where we do want the method to execute
very often was missed.  This patch ensures that setting the config
values low will cause the method to be called as frequently.

Change-Id: I1530c8b3b6ac71f4d9e2e41188c0eca4a6556e01
Closes-Bug: #1274317