Change initial delay for servicegroup api reporting
Initial delay should not necessarily be equal to
report_interval. Instead make it a reasonable value
such as 5 seconds.

This is so the surprising behavior of a service not
reporting when you start it up doesn't happen if your
report interval is configured to a larger value.

For example, in an environment where there are many
services it may not be optimal to have them all reporting
every 10 seconds which is the default. You may have them
report every 5 minutes. It is confusing to start up the
service and not have it show as available until 5 minutes
later.

Fixes bug #1226370

Change-Id: I85c4146bfa465fffaf89d4c2c74e6cb4ef84cee2