Ignore the image name when booting from volume
In 0185d2445 the _run_instance method was changed to acquire the image
metadata from the request_spec instead of glance. This was needed to be
able to get the image_metadata attached to a volume, when booting from
it. That change didn't consider that the image name key is not inherited
when creating a volume from an image and introduced a bug when creating
the usage_info for volume booted instances.

This change fixes this by only attaching the image name to the usage_info
only when booting from an image and not when booting from a volume.

Related-Bug: 1281973
Change-Id: I8ec441e7815855a94a28c4566790d5144deff16f