Add get_hosts to InstanceGroup object
Add a method to the InstanceGroup object that will return a list of
hosts for all non-deleted instances in the group.

Testing this uses a nova.tests.utils method for creating an instance.
A tweak was needed in this test code to ensure that the user_id and
project_id were set on the instance that was created.  Otherwise, this
testing this new method on the InstanceGroup object failed since it
was filtering on the project_id.

Change-Id: Iaf27649ac30b4aa59a3d86f997aa30cfc3d29506