Use instance object instead of _instance_update()
Currently, if a user takes a snapshot of a VM, stops nova-cpu
before the snapshot operation is finished, and restart nova-cpu:
nova-cpu will crash with the error:
AttributeError: 'dict' object has no attribute 'task_state'
This happens because the _instance_update() method returns a dict
and not an Instance object.

This patch fixes the issue by not using _instance_update()
in the compute manager which is returning a dict and using
the instance object instead.

Change-Id: I3af0b09aeff7790b147a684f7c90af17d8ee7ae2
Closes-Bug: #1287945