Mark ESX driver as deprecated
Since the ESX and VC drivers actually share a lot of code, we could
keep the ESX driver.  However, after speaking with some of the VMware
team, the driver could be simplified a good bit if we just drop ESX
support since there is no intention to support it in its current form.

Instead of just pulling it now (based on the hypervisor deprecation
plan that required CI), we'll mark it deprecated for Icehouse and pull
it out for Juno.

DocImpact

Change-Id: I6553e8b39610cae78e5561e8d03a919476f0d99f