xenapi: Fix test_rescue test to ensure assertions are valid
The assertion in this test gathers the list of VDI references and then
checks if the string 'swap' isn't in the list. A VDI reference is a
UUID, so the string 'swap' will never match, even if the code being
tested is buggy.

Compare against an actual VDI reference. Also, add a positive assertion
to check that the root disk from the original VM is attached.

Change-Id: I3e3e9798ce63f2f260b7993620fde3749452bfc9
Closes-bug: 1290903