Revert "enable cloning for rbd-backed ephemeral disks"
This reverts commit c25c60f6a9ab1ccf12f72f76d400e7c9c0d090b3.

The series of patches involved with adding this feature introduced
an unexpected dependency on glance's v2 API, which we do not
currently support. Triggering a user-facing bug quickly, and leaving
some uncertainty about what else is likely to come in the future,
a revert of this code was decided given the short time to -rc1.

Change-Id: Iee835652f56beed1b8ac74d782e34a5238102840
Related-bug: 1291014