Remove the docker driver
The docker compute driver does not currently have CI running against
it.  In accordance with our plan to require this for all drivers, we
are now removing the driver from nova.

https://wiki.openstack.org/wiki/HypervisorSupportMatrix/DeprecationPlan

At this point, the driver should be set up in a stackforge repository
and have CI staged from there to prove its reliability before we could
consider bringing it back into Nova.

DocImpact
UpgradeImpact

Change-Id: I1c9bea2fdeebc4199c4f7d8fca4580a6fb7fed5b