Don't sync [system_]metadata down to cells on instance.save()
A cell should be authoritative for its metadata and system_metadata.
rebuild relis on system_metadata not being updated until the cell
updates it itself.  update_instance_metadata relies on metadata not
being updated so that it can calculate the diff properly.

And before the objects conversion took place the compute.api.update()
method that instance.save() replaced was not used for syncing metadata
or system_metadata.  So this change goes back to previous behaviour for
cells.

Change-Id: I339da9d1ff6d0e59f6a4ab68fd396bdfca7a224b
Closes-bug: #1292181
Closes-bug: #1292185