Make compute API resize methods use Quotas objects
The compute API's various resize methods are using the old-style quota
class to reserve and commit quota changes directly to the database. This
change uses quotas objects to do the reserve and commit instead so those
operations go through conductor for the database access.

Note that the long-term goal is to pass the quota object over RPC
rather than the reservations list, but that's a larger patch which
is not necessary for the bug fix.

Closes-Bug: #1262461

Change-Id: Id18357f449193d25287f32bb8ffbf26ee2194a22