Add deleted_at attribute in glance stub on delete()
The glance stub used in unit tests sets the 'deleted' attribute to
True for the deleted image, but does not also add the 'deleted_at'
attribute, which should also be included, since that's how glance
actually operates. 'deleted_at' is missing from a few other places
as well; add it there for consistency.

Change-Id: I60a96f568d45a4f5fabdb146e0028c4cf680b9fa