always set --no-hosts for dnsmasq
Many network modes in nova-network disable reading host entries from
/etc/hosts. This includes running vish-ha (multihost), or if you ever
set a dns server. However if you don't do these things we continue to
use /etc/hosts.

It's not clear that this does anything other than create
inconsistency. There is a long standing bug that exists around this
because the impacts include exposing localhost as a dns entry on the
network.

Change-Id: Ibccf0999bfd4b37f2d948af34640260f95c54572
Closes-Bug: #1201873