Make obj_make_compatible consistent
The code was originally copy pasted code with an incorrect function
name. It was modified in e274c45dcd37f49d66cc3cce72c2fd97f444151c
to have the proper function names, but no tests were added. The call
is inconsistently used as well. This modifies all the call sites
to work the same and adds a test for one of them.

Change-Id: I79d295ff49057f42f1e6609cecb5bcd38e73f8d9