Ironic driver fetches extra_specs when needed
The Ironic driver needs to access flavor['extra_specs'] in certain
situations. This access was erroneously removed during the large patch
series to land the driver in Nova. The corresponding change did not land
in Ironic, because it fails to pass tempest (as does the current Ironic
driver in Nova).

This patch reverts that change.

Change-Id: I525c611aa0e98076477c2a9a279a0f8a47858aac
Closes-bug: 1367007